chore(mobile): support package types
