+++ apps/mobile/package.json
@@ -83,6 +83,7 @@
     "fast-deep-equal": "^3.1.3",
     "i18next": "^25.2.1",
     "i18next-resources-to-backend": "^1.2.1",
+    "immer": "^10.1.1",
     "jose": "^6.0.8",
     "jsrsasign": "^11.1.0",
     "jwt-decode": "^4.0.0",
@@ -117,7 +118,8 @@
     "ts-md5": "^1.3.1",
     "url-join": "^5.0.0",
     "zod": "^3.24.2",
-    "zustand": "^5.0.3"
+    "zustand": "5.0.4",
+    "zustand-utils": "^2.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.26.10",
