+++ apps/mobile/tsconfig.json
@@ -6,6 +6,7 @@
     "module": "esnext",
     "paths": {
       // incase mixed with parent app
+      "@/types/*": ["../../packages/types/src/*"],
       "@/*": ["./src/*"]
     }
   },
